+++ spec/ultra_settings/configuration_view_spec.rb
@@ -45,7 +45,7 @@ RSpec.describe UltraSettings::ConfigurationView do
     it "renders links for the runtime settings" do
       UltraSettings.runtime_settings_url = "http://example.com/settings?filter=${name}"
       html = UltraSettings::ConfigurationView.new(TestConfiguration.instance).render
-      expect(html).to include('<a href="http://example.com/settings?filter=test.string"')
+      expect(html).to include('href="http://example.com/settings?filter=test.string"')
     end
   end
 end